+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
         <activity
-            android:name=".FavoriteContacts.ScreenSlideActivity"
+            android:name=".components.lastContacts.ScreenSlideActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
