+++ app/src/main/java/naoned/sil/lp/naonedchat/listeners/authentification/DisconnectOnClickListener.java
@@ -1,9 +1,9 @@
-package naoned.sil.lp.naonedchat.authentication.Listeners;
+package naoned.sil.lp.naonedchat.listeners.authentification;
 
 import android.support.design.widget.Snackbar;
 import android.view.View;
 
-import service.Connection;
+import naoned.sil.lp.naonedchat.service.Connection;
 
 /**
  * Created by ACHP on 22/01/2016.
