+++ app/src/main/java/naoned/sil/lp/naonedchat/listeners/chat/myChatListener.java
@@ -1,4 +1,4 @@
-package service;
+package naoned.sil.lp.naonedchat.listeners.chat;
 
 import android.util.Log;
